backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymongo==4.7.2
pydantic-settings==2.4.0
pandas==2.2.2
bottleneck>=1.3.6
numpy==1.26.4
joblib==1.4.2
prophet==1.1.5
statsmodels==0.14.2
scikit-learn==1.5.1
python-multipart==0.0.9
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
black==24.4.2
ruff==0.5.4
mongomock==4.1.2
ortools==9.9.3963
geopy==2.4.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
slowapi==0.1.9
redis==5.0.1
requests==2.31.0
